requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openpyxl==3.1.2
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Limiter==3.5.0
bcrypt==4.1.2
